+++ source/plugins/installationLifeCycle.ts
@@ -4,7 +4,7 @@ import { createInstallation } from "../github/events/create_installation"
 import { RootObject as InstallationCreated } from "../github/events/types/integration_installation_created.types"
 import logger from "../logger"
 
-export const installationLifeCycle = async (event: string, req: express.Request, res: express.Response, ___: any) => {
+export const installationLifeCycle = (event: string, req: express.Request, res: express.Response, ___: any) => {
   if (event !== "installation") {
     return false
   }
@@ -17,7 +17,7 @@ export const installationLifeCycle = async (event: string, req: express.Request,
   if (action === "created") {
     logger.info("")
     logger.info(`## Creating new installation for ${request.installation.account.login}`)
-    await createInstallation(installation, req, res)
+    createInstallation(installation, req, res)
     return true
   }