+++ source/routing/_tests/_router.test.ts
@@ -3,6 +3,16 @@ jest.mock("../../github/events/github_runner", () => ({
   githubDangerRunner: mockGithubRunner,
 }))
 
+jest.mock("../../plugins/installationSettingsUpdater", () => ({
+  installationSettingsUpdater: jest.fn(),
+}))
+
+jest.mock("../../plugins/recordWebhooks", () => ({
+  recordWebhook: jest.fn(),
+}))
+
+import { installationSettingsUpdater } from "../../plugins/installationSettingsUpdater"
+import { recordWebhook } from "../../plugins/recordWebhooks"
 import { githubRouter } from "../router"
 
 const validRequestWithEvent = (header: string) =>
@@ -12,4 +22,13 @@ it("calls the GitHub runner for any events", () => {
   const req = validRequestWithEvent("issue")
   githubRouter(req, {} as any, {} as any)
   expect(mockGithubRunner).toBeCalledWith("issue", req, {}, {})
+})
+
+it("calls plugins", () => {
+  const req = validRequestWithEvent("push")
+  githubRouter(req, {} as any, {} as any)
+
+  expect(mockGithubRunner).toBeCalledWith("push", req, {}, {})
+  expect(installationSettingsUpdater).toBeCalled()
+  expect(recordWebhook).toBeCalled()
 })